Server/requirements.txt
sentence-transformers
groq>=0.5.0
fastapi
uvicorn
pinecone==7.3.0
python-dotenv
